first splitup: raster_colors
