+++ src/actinia_core/processing/common/README.md
@@ -1,0 +1,1 @@
+Shared modules between job receiver and job executor when using redis queue.
