+++ src/actinia_core/rest/raster_colors.py
@@ -4,7 +4,7 @@
 # performance processing of geographical data that uses GRASS GIS for
 # computational tasks. For details, see https://actinia.mundialis.de/
 #
-# Copyright (c) 2016-2018 Sören Gebbert and mundialis GmbH & Co. KG
+# Copyright (c) 2016-2022 Sören Gebbert and mundialis GmbH & Co. KG
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -31,20 +31,17 @@ from flask_restful_swagger_2 import swagger
 from flask import jsonify, make_response
 from actinia_api.swagger2.actinia_core.schemas.raster_colors import RasterColorModel
 
-from .ephemeral_processing import EphemeralProcessing
-from .persistent_processing import PersistentProcessing
-from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
-from tempfile import NamedTemporaryFile
-import os
-import atexit
 from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel, \
     StringListProcessingResultResponseModel
+from actinia_core.rest.base.resource_base import ResourceBase
+from actinia_core.processing.common.raster_colors import start_job_colors_out
+from actinia_core.processing.common.raster_colors import start_job_from_rules
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert, Carmen Tawalika"
-__copyright__ = "Copyright 2016-2021, Sören Gebbert and mundialis GmbH & Co. KG"
+__copyright__ = "Copyright 2016-2022, Sören Gebbert and mundialis GmbH & Co. KG"
 __maintainer__ = "mundialis"
 
 
@@ -210,100 +207,3 @@ class SyncPersistentRasterColorsResource(ResourceBase):
         enqueue_job(self.job_timeout, start_job_from_rules, rdc)
         http_code, response_model = self.wait_until_finish()
         return make_response(jsonify(response_model), http_code)
-
-
-def start_job_colors_out(*args):
-    processing = EphemeralRasterColorsOutput(*args)
-    processing.run()
-
-
-class EphemeralRasterColorsOutput(EphemeralProcessing):
-
-    def __init__(self, *args):
-        EphemeralProcessing.__init__(self, *args)
-
-    def _execute(self, skip_permission_check=True):
-
-        self._setup()
-
-        raster_name = self.map_name
-        self.required_mapsets.append(self.mapset_name)
-
-        with NamedTemporaryFile(
-                mode="w+", suffix=".color", dir=self.temp_file_path) as file:
-            result_file = file.name
-
-        self.request_data = {
-            "1": {
-                "module": "r.colors.out",
-                "inputs": {
-                    "map": "",
-                    "rules": ""
-                }
-            }
-        }
-        self.request_data["1"]["inputs"]["map"] = raster_name + "@" + self.mapset_name
-        self.request_data["1"]["inputs"]["rules"] = result_file
-
-        # Run the selected modules
-        process_list = self._create_temporary_grass_environment_and_process_list(
-            skip_permission_check=True)
-        self._execute_process_list(process_list)
-
-        with open(result_file, "r") as file:
-            result = file.read().strip().split("\n")
-
-        self.module_results = result
-
-
-def start_job_from_rules(*args):
-    processing = PersistentRasterColorsRules(*args)
-    processing.run()
-
-
-class PersistentRasterColorsRules(PersistentProcessing):
-    """Set the color table of a raster map from color rules
-
-    Perform the processing in the original mapset
-    """
-
-    def __init__(self, *args):
-
-        PersistentProcessing.__init__(self, *args)
-
-    def _execute(self, skip_permission_check=True):
-
-        self._setup()
-
-        raster_name = self.map_name
-        self.required_mapsets.append(self.target_mapset_name)
-
-        options = self.request_data
-
-        pc = {"1": {"module": "r.colors", "inputs": {}}}
-        pc["1"]["inputs"]["map"] = raster_name + "@" + self.target_mapset_name
-
-        if "rules" in options:
-            with NamedTemporaryFile(mode="w+", delete=False, suffix=".color",
-                                    dir=self.temp_file_path) as rules:
-                for line in options["rules"]:
-                    rules.write(line + "\n")
-
-            pc["1"]["inputs"]["rules"] = rules.name
-            atexit.register(remove_rules_file, rules.name)
-        else:
-            for option in options:
-                pc["1"]["inputs"][option] = options[option]
-
-        process_list = self._validate_process_chain(process_chain=pc,
-                                                    skip_permission_check=True)
-        self._check_lock_target_mapset()
-        self._create_temp_database(self.required_mapsets)
-        self._create_grass_environment(grass_data_base=self.temp_grass_data_base,
-                                       mapset_name=self.target_mapset_name)
-
-        self._execute_process_list(process_list)
-
-
-def remove_rules_file(rules_file):
-    os.remove(rules_file)
